requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
django-filter==23.5
Pillow==10.2.0
djangorestframework-gis==1.0 